Found the function giving errors
LiveMatch needs to be passed in an opponent's username, or null, to know whether it's creating or joining a game. the function in CreateGameBox to find an existing user opens a stream that continually throws exceptions until the server is restarted. On the bright side, the postGameRequest method seems to work fine.