Cleaned up CreateGame and JoinGame CSS
CreateGame and JoinGame both displayed things that we hope to support one day, but will not make it to the final release. This includes correspondence games and changing the join table's look. The "filter by" dropdown on the JoinGame component is not needed anymore, since the table's column headers have the same functionality.